refactor TTTR merge: enhance memory handling and conditional allocation
- Updated memory allocation to use HDF5-specific allocator for `PHOTON_HDF_CONTAINER`.
- Introduced error handling for HDF5 builds without required support.
- Replaced raw memory deallocation with `deallocate_memory_of_records` for consistency.
- Adjusted `merge` method parameters for type consistency and clarity.

+++ include/TTTR.h
@@ -788,8 +788,8 @@ public:
 
     void merge(
         const TTTR& other, 
-        uint64_t offset_macro_time, 
-        int32_t channel_offset, 
+        unsigned long long offset_macro_time, 
+        int channel_offset, 
         int strategy = 0
     );
